Refactor kubemodel compute to core
